+++ packages/widgets/role-management-widget/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.8.0](https://github.com/descope/descope-js/compare/role-management-widget-0.7.2...role-management-widget-0.8.0) (2026-06-09)
+
+### Dependency Updates
+
+* `sdk-mixins` updated to version `0.20.0`
+
+### Features
+
+* **web-component:** client-side realtime evaluation of componentsConditions RELEASE ([#1404](https://github.com/descope/descope-js/issues/1404)) ([adab0d9](https://github.com/descope/descope-js/commit/adab0d9ef33affb83c919af3f4d95759f89a6ae5)), closes [descope/etc#14360](https://github.com/descope/etc/issues/14360)
+
 ## [0.7.2](https://github.com/descope/descope-js/compare/role-management-widget-0.7.1...role-management-widget-0.7.2) (2026-06-07)
 
 ### Dependency Updates